requirements.txt
Pillow
paramiko
